- DBViewModel deleted
