--- app/src/main/java/co/develhope/chooseyouowncocktail_g2/MainViewModel.kt
@@ -1,16 +1,0 @@
-package co.develhope.chooseyouowncocktail_g2
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import co.develhope.chooseyouowncocktail_g2.domain.DBViewModel
-import co.develhope.chooseyouowncocktail_g2.network.DrinksProvider
-
-
-class MainViewModel : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(DBViewModel::class.java)) {
-            return DBViewModel() as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
