+++ app/src/main/java/co/develhope/chooseyouowncocktail_g2/ui/home/HomeFragment.kt
@@ -1,6 +1,5 @@
 package co.develhope.chooseyouowncocktail_g2.ui.home
 
-import android.opengl.Visibility
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -27,8 +26,6 @@ import co.develhope.chooseyouowncocktail_g2.databinding.FragmentHomeBinding
 import co.develhope.chooseyouowncocktail_g2.domain.DBEvent
 import co.develhope.chooseyouowncocktail_g2.domain.DBResult
 import co.develhope.chooseyouowncocktail_g2.domain.DBViewModel
-
-import com.google.android.material.snackbar.Snackbar
 
 
 import co.develhope.chooseyouowncocktail_g2.ui.DetailDrinkFragment
